+++ dialect/mssqldialect/dialect.go
@@ -51,7 +51,7 @@ func New() *Dialect {
 	return d
 }
 
-func (d *Dialect) Init(db bun.SQLRepo) {
+func (d *Dialect) Init(db schema.SQLRepo) {
 	var version string
 	if err := db.QueryRow("SELECT @@VERSION").Scan(&version); err != nil {
 		log.Printf("can't discover MSSQL version: %s", err)
